Removed file-processor.ts and added comments to file-renderer.ts
