+++ src/doc-renderer.ts
@@ -1,12 +1,12 @@
-import Renderer from "./renderer/renderer";
+import FileRenderer from "./renderer/file-renderer";
 import {MarkdownRenderer} from "./renderer/markdown-renderer";
 import {LiquidRenderer} from "./renderer/liquid-renderer";
 
-export class FileRenderer implements Renderer {
+export class DocRenderer {
 
   private readonly data: Object;
-  private markdown: Renderer;
-  private liquid: Renderer;
+  private markdown: FileRenderer;
+  private liquid: FileRenderer;
 
   /**
    * Constructor that takes the data
@@ -26,11 +26,11 @@ export class FileRenderer implements Renderer {
     return this.liquid.render(content, {...this.data, ...data});
   }
 
-  public async renderFile(path: string, data: Object = {}): Promise<string> {
+  public async renderMarkdownFile(path: string, data: Object = {}): Promise<string> {
     return this.markdown.renderFile(path, {...this.data, ...data});
   }
 
-  public async render(content: string, data: Object = {}): Promise<string> {
+  public async renderMarkdownTemplate(content: string, data: Object = {}): Promise<string> {
     const contents = await this.renderLiquidTemplate(content, {...this.data, ...data});
     return this.markdown.render(contents, {});
   }
