+++ src/renderer/liquid-renderer.ts
@@ -1,8 +1,8 @@
-import Renderer from "./renderer";
+import FileRenderer from "./file-renderer";
 import {Liquid} from "liquidjs";
 import {readFile} from "../utils/file-utils";
 
-export class LiquidRenderer implements Renderer {
+export class LiquidRenderer implements FileRenderer {
 
   private liquid: Liquid;
