--- src/renderer/renderer.ts
@@ -1,6 +1,0 @@
-export default interface Renderer {
-  renderFile(path: string, data: Object)
-
-  render(content: string, data: Object);
-}
-
