+++ src/utils/file-utils.ts
@@ -16,4 +16,21 @@ export const writeFile = promisify(fs.writeFile);
  */
 export const copyFile = promisify(fs.copyFile);
 
+/**
+ * Make a directory in the filesystem.
+ */
 export const mkdir = promisify(fs.mkdir);
+
+/**
+ * Take a list of strings and join the file contents of their files.
+ *
+ * @param files to join the contents of
+ */
+export const joinFiles = async (files: string[]): Promise<string> => {
+  let result = '';
+  for (const value of files) {
+    let content = (await readFile(value)).toString();
+    result += content;
+  }
+  return result;
+}
